refactor: split pipeline issues.js and visualStages.js along CRUD/stage/sync seams (#2531)
